Update of the configure+build setup.

+++ autogen.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#! /bin/sh
 
 # Regenerate all files which are constructed by the autoconf, automake
 # and libtool tool-chain. Note: only developers should need to use
@@ -9,15 +9,23 @@
 #   Lars J. Aas <larsa@sim.no>
 #   Ralph Levien (original autogen.sh picked up from Gnome source archive).
 
-DIE=false
+directory=`echo "$0" | sed -e 's/[^\/]*$//g'`;
+cd $directory
+if ! test -f ./autogen.sh; then
+  echo "unexpected problem with your shell - bailing out"
+  exit 1
+fi
 
 GUI=Xt
 
 PROJECT=So$GUI
 MACRODIR=conf-macros
-SUBPROJECTS="$MACRODIR examples"
+DIE=false
 
+SUBPROJECTS="$MACRODIR data examples"
+SUBPROJECTNAMES="$MACRODIR SoXtData SoXtExamples"
 AUTOMAKE_ADD=
+
 if test "$1" = "--clean"; then
   rm -f aclocal.m4 \
 	config.guess \
@@ -81,20 +89,29 @@ if test -z "`libtool --version | grep \" $LIBTOOL_VER \" 2> /dev/null`"; then
     DIE=true
 fi
 
+set $SUBPROJECTNAMES
+num=1
 for project in $SUBPROJECTS; do
   test -d $project || {
-    echo
-    echo "The CVS sub-project '$project' was not found."
-    echo "It was probably added after you initially checked out $PROJECT."
-    echo "Do a fresh 'cvs checkout' to correct this problem - the $PROJECT build system"
-    echo "will probably not work properly otherwise.  For a fresh 'cvs checkout',"
-    echo "run 'cvs -d :pserver:cvs@cvs.sim.no:/export/cvsroot co -P $PROJECT'."
-    echo
+    echo "Could not find subdirectory '$project'."
+    echo "It was probably added after you initially fetched $PROJECT."
+    echo "To add the missing module, run 'cvs co $1' from the $PROJECT"
+    echo "base directory."
+    echo ""
+    echo "To do a completely fresh cvs checkout of the whole $PROJECT module,"
+    echo "(if all else fails), remove $PROJECT and run:"
+    echo ""
+    echo "  cvs -z3 -d :pserver:cvs@cvs.sim.no:/export/cvsroot co -P $PROJECT"
+    echo ""
+    DIE=true
   }
+  num=`expr $num + 1`
+  shift
 done
 
 $DIE && exit 1
 
+# AG_CRUNCH
 echo "Running aclocal (generating aclocal.m4)..."
 aclocal -I $MACRODIR
 
@@ -105,7 +122,7 @@ echo "Running automake (generating the Makefile.in files)..."
 echo "[ignore any \"directory should not contain '/'\" warning]"
 automake $AUTOMAKE_ADD
 
-AMBUGFIXES=`find . \( -name Makefile.in.diff \) | egrep -v '^\./(examples|ivexamples)'`
+AMBUGFIXES=`find . \( -name Makefile.in.diff \) | egrep -v '^\./(data|examples|ivexamples)'`
 
 fixmsg=0
 for bugfix in $AMBUGFIXES; do
